Istio and reservation check bug fix
