+++ lib/global-admin/addon/components/registry-catalog/component.js
@@ -580,20 +580,7 @@ export default Component.extend(CrudCatalog, ReservationCheck, {
         notaryNodeSelector[key] = value
       });
 
-      const workloads = this.getEnalbedWorkloads(answers)
-
-
-      const preRequestsCpu = workloads.reduce((all, current) => {
-        const value = answers[`${ current }.resources.requests.cpu`]
-
-        return value ? all + convertToMillis(value) : all
-      }, 0)
-
-      const preRequestsMemory = workloads.reduce((all, current) => {
-        const value = answers[`${ current }.resources.requests.memory`]
-
-        return value ? all + parseSi(value) / 1048576 : all
-      }, 0)
+      this.updatePreRequest();
 
       setProperties(this, {
         registryNodeSelector,
@@ -601,8 +588,6 @@ export default Component.extend(CrudCatalog, ReservationCheck, {
         redisNodeSelector,
         clairNodeSelector,
         notaryNodeSelector,
-        preRequestsCpu,
-        preRequestsMemory,
       })
 
       let persistenceKeys = []
@@ -796,7 +781,8 @@ export default Component.extend(CrudCatalog, ReservationCheck, {
   },
 
   doneSaving() {
-    this.setGlobalRegistryEnabled('true')
+    this.setGlobalRegistryEnabled('true');
+    this.updatePreRequest();
   },
 
   setGlobalRegistryEnabled(value, successCB) {
