+++ lib/istio/addon/components/istio-catalog/component.js
@@ -643,19 +643,7 @@ export default Component.extend(CrudCatalog, ReservationCheck, {
       loadBalancerSourceRanges.pushObject(answers[k])
     })
 
-    const workloads = this.getEnalbedWorkloads(answers)
-
-    const preRequestsCpu = workloads.reduce((all, current) => {
-      const value = answers[`${ current }.resources.requests.cpu`]
-
-      return value ? all + convertToMillis(value) : all
-    }, 0)
-
-    const preRequestsMemory = workloads.reduce((all, current) => {
-      const value = answers[`${ current }.resources.requests.memory`]
-
-      return value ? all + parseSi(value) / 1048576 : all
-    }, 0)
+    this.updatePreRequest()
 
     setProperties(this, {
       prometheusNodeSelector,
@@ -667,8 +655,6 @@ export default Component.extend(CrudCatalog, ReservationCheck, {
       loadBalancerSourceRanges,
       'config.http2Port':         answers[HTTP2_PORT] || DEFAULT_HTTP2_PORT,
       'config.httpsPort':         answers[HTTPS_PORT] || DEFAULT_HTTPS_PORT,
-      preRequestsCpu,
-      preRequestsMemory,
     })
 
     Object.keys(answers).forEach((key = '') => {
@@ -772,5 +758,9 @@ export default Component.extend(CrudCatalog, ReservationCheck, {
     }
 
     return out;
-  }
+  },
+
+  doneSaving() {
+    this.updatePreRequest()
+  },
 });
