+++ lib/istio/translations/en-us.yaml
@@ -31,13 +31,6 @@ clusterIstioPage:
       label: Enable Persistent Storage for Grafana
     mtlsEnabled:
       label: Service to service mutual TLS
-    pilotCpu:
-      label: Pilot CPU Reservation
-      placeholder: e.g. 500
-      unit: milli CPUs
-    pilotMemory:
-      label: Pilot Memory Reservation
-      placeholder: e.g. 2048
     gatewayEnabled:
       label: Enabled Ingress Gateway
     gatewayType:
@@ -52,29 +45,11 @@ clusterIstioPage:
     prometheusRetention:
       label: Retention for Prometheus
       placeholder: e.g. 6h
-    prometheusRequestMemory:
-      label: Prometheus Memory Reservation
-      placeholder: e.g. 1000
-    prometheusLimitMemory:
-      label: Prometheus Memory Limit
-      placeholder: e.g. 1000
-    prometheusLimitCpu:
-      label: Prometheus CPU Limit
-      placeholder: e.g. 1000
-    prometheusRequestCpu:
-      label: Prometheus CPU Reservation
-      placeholder: e.g. 1000
     traceSampling:
       label: Trace sampling Percentage
       placeholder: e.g. 1
     mixerPolicyEnabled:
       label: Enabled Mixer Policy
-    pilotLimitCpu:
-      label: Pilot CPU Limit
-      placeholder: e.g. 1000
-    pilotLimitMemory:
-      label: Pilot Memory Limit
-      placeholder: e.g. 1000
   nodeSelector:
     addSelectorLabel: Add Selector
     helpText: Select the nodes where {component} workloads will be scheduled to
