+++ lib/shared/addon/mixins/reservation-check.js
@@ -1,6 +1,6 @@
 import { inject as service } from '@ember/service';
 import Mixin from '@ember/object/mixin';
-import { get, computed } from '@ember/object';
+import { get, computed, setProperties } from '@ember/object';
 import { convertToMillis, ucFirst, parseCamelcase } from 'shared/utils/util';
 import { parseSi } from 'shared/utils/parse-unit';
 import { requiredError } from 'shared/utils/util';
@@ -31,7 +31,7 @@ export default Mixin.create({
     let leftCpu = 0;
 
     schedulableNodes.forEach((node) => {
-      const left =  convertToMillis(get(node, 'allocatable.cpu')) - convertToMillis(get(node, 'requested.cpu') || '0');
+      const left =  convertToMillis(get(node, 'allocatable.cpu') || '0') - convertToMillis(get(node, 'requested.cpu') || '0');
 
       leftCpu += left;
     });
@@ -196,5 +196,27 @@ export default Mixin.create({
     })
 
     return errors
+  },
+
+  updatePreRequest() {
+    const answers = get(this, 'app.answers') || {};
+    const workloads = this.getEnalbedWorkloads(answers);
+
+    const preRequestsCpu = workloads.reduce((all, current) => {
+      const value = answers[`${ current }.resources.requests.cpu`]
+
+      return value ? all + convertToMillis(value) : all
+    }, 0)
+
+    const preRequestsMemory = workloads.reduce((all, current) => {
+      const value = answers[`${ current }.resources.requests.memory`]
+
+      return value ? all + parseSi(value) / 1048576 : all
+    }, 0)
+
+    setProperties(this, {
+      preRequestsCpu,
+      preRequestsMemory,
+    })
   },
 });
